Add account dashboard application (#32)
* Moved dashboard to monorepo

* Fixed submodule issue

* Successfully added dashboard

* Update module resolution strategy in tsconfig

The module resolution strategy in the TypeScript configuration for the Abstraxion Dashboard app has been updated. Previously, it was set to "bundler", but has now been changed to "node" fixing an known
issue with the abstraxion package.

* Add initial commit of dashboard

A new file has been created for the initial setup of the dashboard. This forms part of the foundation for the "abstraxion-dashboard".

* Moved abstraxion to dashboard

* Resolved build errors

---------

Co-authored-by: Burnt Nerve <seanlbray@gmail.com>