Add more grant types which can be requested from a dapp (#107)
* Add support for bank grants

This update enhances the Abstraxion dashboard by adding support for bank grants. It now handles "bank" as a query parameter, thus allowing for the specification of bandwidth spending limits. The commit also entails modifying the AbstraxionProvider and AbstraxionSignin components to accommodate this feature. Additional helper functions for generating stake, bank and contract grants were introduced.

* Add staking capability to Abstraxion components

Added a "stake" boolean configuration option to the Abstraxion, AbstraxionContext and AbstraxionSignin components. When set to true, stake-related authorization grants are generated in addition to contract grants in the AbstraxionGrant component. This allows users to stake assets using the Abstraxion interface.